Fix pandas deprecation warnings in timestamp conversion

+++ crypto_ma_monitor.py
@@ -41,7 +41,7 @@ def get_okx_kline_data(symbol, interval=KLINE_INTERVAL, limit=KLINE_LIMIT):
         ])
         
         # 转换数据类型
-        df['timestamp'] = pd.to_datetime(df['timestamp'], unit='ms')
+        df['timestamp'] = pd.to_datetime(df['timestamp'].astype(float), unit='ms')
         df['close'] = df['close'].astype(float)
         df['open'] = df['open'].astype(float)
         df['high'] = df['high'].astype(float)
@@ -73,7 +73,7 @@ def get_binance_kline_data(symbol, interval=KLINE_INTERVAL, limit=KLINE_LIMIT):
         ])
         
         # 转换数据类型
-        df['timestamp'] = pd.to_datetime(df['timestamp'], unit='ms')
+        df['timestamp'] = pd.to_datetime(df['timestamp'].astype(float), unit='ms')
         df['close'] = df['close'].astype(float)
         df['open'] = df['open'].astype(float)
         df['high'] = df['high'].astype(float)
